feat: 서비스 테스트 모두 삭제 컨트롤러 테스트 모두 삭제 infra 테스트 모두 삭제 City 및 District 레파지토리 테스트 재정의 및 테스트 완료 application-test.yml 수정
