+++ src/test/java/com/jangmo/web/repository/CityRepositoryTest.java
@@ -1,52 +1,41 @@
 package com.jangmo.web.repository;
 
-import com.jangmo.web.constants.message.ErrorMessage;
-import com.jangmo.web.exception.NotFoundException;
+import com.jangmo.web.config.JpaConfiguration;
 import com.jangmo.web.model.entity.administrative.City;
-
-import org.junit.jupiter.api.DisplayName;
+import com.jangmo.web.model.entity.administrative.District;
+import lombok.extern.slf4j.Slf4j;
+import org.checkerframework.checker.units.qual.C;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.jdbc.Sql;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.ActiveProfiles;
 
-import java.util.Optional;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-@Sql("/test_sql/area_city_test.sql")
+@Slf4j
 @DataJpaTest
+@ActiveProfiles("test")
+@Import(JpaConfiguration.class)
 public class CityRepositoryTest {
 
     @Autowired
     private CityRepository cityRepository;
 
+    @Autowired
+    private DistrictRepository districtRepository;
 
-    @DisplayName("city 데이터 확인 테스트")
+    static final String CITY_NAME = "서울특별시";
+    static final String DISTRICT_NAME = "강남구";
+
     @Test
-    void getCityInfoTest() {
-        String cityName = "서울특별시";
-        Optional<City> seoul = cityRepository.findById(1L);
-        assertThat(seoul).isPresent();
-        assertThat(seoul.get().getName()).isEqualTo(cityName);
+    void 시도_저장_테스트() {
+        City city = City.of(CITY_NAME);
+        cityRepository.save(city);
+        assertNotNull(city.getId());
+        log.info("city id : {}", city.getId());
     }
 
-    @DisplayName("city 데이터 not found 테스트")
-    @Test
-    void getCityNotFoundException() {
-        String cityName = "없는 도시명 테스트";
-        City city = cityRepository.findByName(cityName).orElseThrow(
-                () -> new NotFoundException(ErrorMessage.CITY_NOT_FOUND)
-        );
-    }
-
-    @DisplayName("city 카운팅 테스트")
-    @Test
-    void getCityCountTest() {
-        int count = 17;
-        int savedCount = cityRepository.findAll().size();
-        assertThat(savedCount).isEqualTo(count);
-    }
 
 }
