+++ src/test/java/com/jangmo/web/repository/DistrictRepositoryTest.java
@@ -1,20 +1,22 @@
 package com.jangmo.web.repository;
 
+import com.jangmo.web.config.JpaConfiguration;
 import com.jangmo.web.model.entity.administrative.City;
 import com.jangmo.web.model.entity.administrative.District;
-import org.junit.jupiter.api.DisplayName;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.jdbc.Sql;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.ActiveProfiles;
 
-import java.util.List;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-@Sql("/test_sql/area_city_test.sql")
-@Sql("/test_sql/area_district_test.sql")
+@Slf4j
 @DataJpaTest
+@ActiveProfiles("test")
+@Import(JpaConfiguration.class)
 public class DistrictRepositoryTest {
 
     @Autowired
@@ -23,22 +25,23 @@ public class DistrictRepositoryTest {
     @Autowired
     private DistrictRepository districtRepository;
 
-    @DisplayName("district 데이터 확인 테스트")
-    @Test
-    void getDistrictInfoTest() {
-        City seoul = cityRepository.findByName("서울특별시").get();
-        String districtName = "강남구";
-        List<District> gangNam = districtRepository.findByCity(seoul);
-        assertThat(gangNam.get(22).getCity()).isEqualTo(seoul);
-        assertThat(gangNam.get(22).getName()).isEqualTo(districtName);
+    static final String CITY_NAME = "서울특별시";
+    static final String DISTRICT_NAME = "강남구";
+    City city;
+    @BeforeEach
+    void setup() {
+        city = City.of(CITY_NAME);
+        cityRepository.save(city);
     }
 
-    @DisplayName("district 카운팅 테스트")
+
     @Test
-    void getDistrictCountTest() {
-        int count = 229;
-        int savedCount = districtRepository.findAll().size();
-        assertThat(savedCount).isEqualTo(count);
+    void 시군구_저장_테스트() {
+        District district = District.of(DISTRICT_NAME, city);
+        districtRepository.save(district);
+
+        assertNotNull(district.getId());
+        log.info("district id : {}", district.getId());
     }
 
 }
